media table changed

+++ src/main/resources/db/scripts/V8__MEDIA_TABLE.sql
@@ -4,6 +4,6 @@ create table media
     media_type  TINYINT      NOT NULL,
     media_url   VARCHAR(255) NOT NULL,
     thumbnail   boolean default false,
-    sequence_id TINYINT      NOT NULL,
+    sequence_id BIGINT      NOT NULL,
     item_id     BIGINT       NOT NULL
 );
